requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncio-mqtt==0.16.1
paho-mqtt==1.6.1
